+++ datahub_common/lib/common.dart
@@ -1,8 +1,9 @@
 library cl_datahub_common;
 
 export 'src/api/dto/encoding.dart';
-export 'src/api/dto/field.dart';
 export 'src/api/dto/transfer_object.dart';
+export 'src/api/dto/transfer_object_base.dart';
+export 'src/api/dto/transfer_bean.dart';
 
 export 'src/utils/api_error.dart';
 export 'src/utils/api_exception.dart';