+++ datahub_common/lib/src/api/dto/encoding.dart
@@ -1,20 +1,16 @@
 import 'dart:convert';
 import 'dart:typed_data';
+import 'package:boost/boost.dart';
 
 import 'package:cl_datahub_common/common.dart';
 
-T? decodeTyped<T>(dynamic raw, {DTOFactory? factory}) {
+T decodeTyped<T>(dynamic raw) {
+  return decodeTypedNullable(raw) ?? (throw ApiException('Missing value.'));
+}
+
+T? decodeTypedNullable<T>(dynamic raw) {
   if (raw == null) {
     return null;
-  }
-
-  if (factory != null) {
-    final result = factory(raw);
-    if (result is! T) {
-      throw ApiError('Factory returned wrong type: $result (should be $T)');
-    }
-
-    return result;
   }
 
   if (raw is T) {
@@ -69,9 +65,62 @@ dynamic encodeTyped<T>(T? value) {
     return value;
   }
 
-  if (value is TransferObject) {
-    return value.toJson();
-  }
-
   throw ApiError.invalidType(T);
 }
+
+List<T> decodeList<T>(dynamic raw, T Function(dynamic) decoder) {
+  if (raw is List) {
+    return raw.whereNotNull.map(decoder).toList();
+  } else {
+    return [];
+  }
+}
+
+List encodeList<T>(List<T>? value, dynamic Function(T) encoder) {
+  if (value == null) {
+    return [];
+  }
+
+  return value.map(encoder).toList();
+}
+
+Map<String, V> decodeStringMap<V>(dynamic raw, V Function(dynamic) decoder) {
+  if (raw is Map<String, dynamic>) {
+    return raw.map((key, value) => MapEntry(key, decoder(value)));
+  } else {
+    return {};
+  }
+}
+
+Map<String, dynamic> encodeStringMap<V>(
+    Map<String, V>? value, dynamic Function(V) encoder) {
+  if (value == null) {
+    return {};
+  }
+
+  return value.map((key, value) => MapEntry(key, encoder(value)));
+}
+
+T decodeEnum<T>(dynamic raw, List<T> values) {
+  if (raw is String) {
+    return findEnum(raw, values);
+  } else if (raw is T) {
+    return raw;
+  } else {
+    throw ApiError.invalidType(T);
+  }
+}
+
+T? decodeEnumNullable<T>(dynamic raw, List<T> values) {
+  if (raw is String) {
+    return findEnum(raw, values);
+  } else if (raw is T) {
+    return raw;
+  } else if (raw == null) {
+    return null;
+  } else {
+    throw ApiError.invalidType(T);
+  }
+}
+
+String encodeEnum<T>(T value) => enumName<T>(value);