+++ datahub_common/lib/src/api/dto/transfer_object_base.dart
@@ -1,0 +1,9 @@
+import 'transfer_bean.dart';
+
+/// Interface used internally to detect transfer objects.
+///
+/// DO NOT implement directly, extend generated _TransferObject instead.
+abstract class TransferObjectBase {
+  dynamic toJson();
+  TransferBean get bean;
+}