+++ lib/src/utils/utils.dart
@@ -1,5 +1,3 @@
-import 'package:cl_datahub/cl_datahub.dart';
-
 const Map<int, String> _statusCodes = {
   // Informative
   100: 'Continue',
@@ -81,16 +79,4 @@ String buildQueryString(Map<String, String> query) {
           .map((e) =>
               '${Uri.encodeQueryComponent(e.key)}=${Uri.encodeQueryComponent(e.value)}')
           .join('&');
-}
-
-dynamic customJsonEncode(dynamic item) {
-  if (item is DateTime) {
-    return item.toIso8601String();
-  }
-
-  if (item is TransferObject) {
-    return item.toJson();
-  }
-
-  return item;
 }